Build overhaul
This repository is now dedicated only to nightly builds and unstable
applications.

Stable builds are provided by Flathub.

If an app is not yet available on Flathub it means that it's not yet
compatible with Flatpak.

Builds are now performed with GitHub workflows and are uploaded to the
Liri server using flat-manager.

We also update the manifest files to the new CMake build system that we
are using now instead of qbs.